+++ src/main/java/lazyae2/block/BlockMachine.java
@@ -37,6 +37,7 @@ import lazyae2.LazyAE2;
 import lazyae2.core.LazyAE2Config;
 import lazyae2.core.LazyAE2Tab;
 import lazyae2.tile.TileAggregator;
+import lazyae2.tile.TileCentrifuge;
 import lazyae2.tile.TileProcessor;
 import appeng.util.Platform;
 
@@ -56,7 +57,8 @@ public final class BlockMachine extends Block {
      */
     public enum Type implements IStringSerializable {
 
-        AGGREGATOR(TileAggregator::new);
+        AGGREGATOR(TileAggregator::new),
+        CENTRIFUGE(TileCentrifuge::new);
 
         private static final Type[] VALUES = values();